# api.yaml
openapi: 3.1.0
info:
  title: MikuPush Server API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
  - url: https://mikupush.io
  - url: http://localhost:8080
paths:
  /api/file:
    post:
      summary: Register new file upload
      description: Register new file upload
      operationId: registerFile
      requestBody:
        description: Register new file upload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileCreate'
        required: true
      responses:
        200:
          description: Successful operation
        409:
          description: File registration exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        400:
          description: Request body is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        413:
          description: File exceeds the size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/file/{id}:
    get:
      summary: Get registered file info
      description: Get registered file info
      operationId: getFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: acdb1260-d4de-4aab-a3a9-02665ae25771
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileInfo'
        400:
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: File upload not exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a file upload
      description: Delete a file upload
      operationId: deleteFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: acdb1260-d4de-4aab-a3a9-02665ae25771
      responses:
        200:
          description: Successful operation
        400:
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: File upload not exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/file/{id}/upload:
    post:
      summary: Upload file content for registered file upload
      description: Upload file content for registered file upload
      operationId: uploadFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: acdb1260-d4de-4aab-a3a9-02665ae25771
      requestBody:
        description: The file contents
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
        required: true
      responses:
        200:
          description: Successful operation
        400:
          description: Invalid request parameters or upload is not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        413:
          description: File contents exceeds the size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: File is not registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /health:
    get:
      summary: Health check
      description: |
        Health check endpoint that returns status in different formats
        based on the Accept header:
        - `application/json`: Returns JSON with health status
        - `text/html`: Returns HTML page with health status

        For now it only check the database connection executing the 'SELECT 1' query.
      parameters:
        - name: Accept
          in: header
          description: Preferred content type for the response
          required: false
          schema:
            type: string
            enum:
              - "application/json"
              - "text/html"
            default: "text/html"
      responses:
        '200':
          description: Service health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
            example:
              status: up
            text/html:
              schema:
                type: string
                format: html
              description: Healthy service HTML page
        '500':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              example:
                status: down
            text/html:
              schema:
                type: string
                format: html
              description: Unhealthy status HTML page

components:
  schemas:
    Error:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    FileCreate:
      required:
        - id
        - name
        - mime_type
        - size
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: acdb1260-d4de-4aab-a3a9-02665ae25771
        name:
          type: string
          example: dogs.jpeg
        mime_type:
          type: string
          example: image/jpeg
        size:
          type: integer
          format: int64
          example: 3145728
    FileInfo:
      required:
        - id
        - name
        - mime_type
        - size
        - status
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: acdb1260-d4de-4aab-a3a9-02665ae25771
        name:
          type: string
          example: dogs.jpeg
        mime_type:
          type: string
          example: image/jpeg
        size:
          type: integer
          format: int64
          example: 3145728
        uploaded_at:
          type: string
          format: date-time
          example: "2021-09-29T12:00:00Z"
        status:
          type: string
          enum:
            - WaitingForUpload
            - Uploaded
    HealthStatus:
      required:
        - status
      type: object
      properties:
        status:
          type: string
          enum:
            - "up"
            - "down"